requirements.txt
functions-framework==3.*
google-cloud-spanner==3.*
langchain-google-vertexai==0.*
requests==2.*
scikit-learn==1.*
